+++ integration-tests/modules/__tests__/order/order.spec.ts
@@ -20,7 +20,7 @@ const env = {}
 
 medusaIntegrationTestRunner({
   env,
-  testSuite: ({ dbConnection, getContainer, api }) => {
+  testSuite: ({ dbConnection, getContainer, api, dbUtils }) => {
     let appContainer
     let orderModule: IOrderModuleService
 
@@ -29,8 +29,10 @@ medusaIntegrationTestRunner({
       orderModule = appContainer.resolve(Modules.ORDER)
     })
 
-    beforeEach(async () => {
+    beforeAll(async () => {
       await createAdminUser(dbConnection, adminHeaders, appContainer)
+
+      await dbUtils.snapshot()
     })
 
     describe("CreateOrderWorkflow", () => {
@@ -280,7 +282,7 @@ medusaIntegrationTestRunner({
           id: expect.any(String),
           status: "pending",
           version: 1,
-          display_id: 2,
+          display_id: created.display_id,
           custom_display_id: null,
           payment_collections: [],
           payment_status: "not_paid",
@@ -595,6 +597,8 @@ medusaIntegrationTestRunner({
                     precision: 20,
                   },
                   provider_id: null,
+                  data: null,
+                  metadata: null,
                   created_at: expect.any(String),
                   updated_at: expect.any(String),
                   deleted_at: null,