+++ integration-tests/modules/__tests__/product/admin/products.spec.ts
@@ -36,7 +36,7 @@ async function createProductsWithVariants(
 
 medusaIntegrationTestRunner({
   env,
-  testSuite: ({ dbConnection, getContainer, api }) => {
+  testSuite: ({ dbConnection, getContainer, api, dbUtils }) => {
     describe("Admin: Products API", () => {
       let appContainer
       let product
@@ -54,8 +54,10 @@ medusaIntegrationTestRunner({
         productModule = appContainer.resolve(Modules.PRODUCT)
       })
 
-      beforeEach(async () => {
+      beforeAll(async () => {
         await createAdminUser(dbConnection, adminHeaders, appContainer)
+
+        await dbUtils.snapshot()
       })
 
       describe("GET /admin/products", () => {