+++ packages/admin/dashboard/src/routes/roles/role-list/role-list.tsx
@@ -1,10 +1,10 @@
-import { SingleColumnPage } from "../../../components/layout/pages"
+import { CORE_LAYOUT_IDS } from "@medusajs/admin-shared"
+
+import { LayoutComposer } from "../../../components/layout-composer"
 import { useRequireRbacFeature } from "../../../hooks/use-require-rbac-feature"
-import { useExtension } from "../../../providers/extension-provider"
 import { RoleListTable } from "./components/role-list-table"
 
 export const RoleList = () => {
-  const { getWidgets } = useExtension()
   const isRbacEnabled = useRequireRbacFeature()
 
   if (!isRbacEnabled) {
@@ -12,13 +12,16 @@ export const RoleList = () => {
   }
 
   return (
-    <SingleColumnPage
-      widgets={{
-        before: getWidgets("role.list.before"),
-        after: getWidgets("role.list.after"),
+    <LayoutComposer
+      widgetsZonePrefix="role.list"
+      preferredLayoutId={CORE_LAYOUT_IDS.SINGLE_COLUMN}
+      sections={{
+        main: (
+          <LayoutComposer.Entry id="RoleListTable">
+            <RoleListTable />
+          </LayoutComposer.Entry>
+        ),
       }}
-    >
-      <RoleListTable />
-    </SingleColumnPage>
+    />
   )
 }