+++ www/apps/api-reference/specs/store/code_samples/Shell/store_product-options_{id}/get.sh
@@ -1,0 +1,2 @@
+curl '{backend_url}/store/product-options/{id}' \
+-H 'x-publishable-api-key: {your_publishable_api_key}'