+++ www/packages/tags/src/tags/how-to.ts
@@ -20,6 +20,10 @@ export const howTo = [
     "path": "https://docs.medusajs.com/resources/commerce-modules/order/order-totals"
   },
   {
+    "title": "Filter Products",
+    "path": "https://docs.medusajs.com/resources/commerce-modules/product/guides/filter-products"
+  },
+  {
     "title": "Get Variant Prices",
     "path": "https://docs.medusajs.com/resources/commerce-modules/product/guides/price"
   },